pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mock_ai-agent"
version = "0.1.0"
description = "Agente conversacional de LangGraph para agendación de citas"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.4",
    "langgraph-cli[inmem]>=0.1.0",
    "langchain-openai>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "langchain-google-genai>=3.2.0",
    "langchain-core>=1.1.0",
    "python-dotenv>=1.2.1",
    "pydantic>=2.0.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=9.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
